+++ apps/remnant2toolkit/app/(features)/_genesis-code/_components/genesis-code-display.tsx
@@ -21,6 +21,10 @@ export function GenesisCodeDisplay() {
       <h2 className="text-surface-solid w-full text-center text-2xl font-bold">
         Genesis Code
       </h2>
+      <div className="bold w-full text-center text-xl text-red-500">
+        Genesis code is currently not calculating correctly. We are working on a
+        fix. (2025-09-16)
+      </div>
       <div className="flex w-full flex-col items-center justify-center">
         <div
           id="code-container"
@@ -53,7 +57,7 @@ export function GenesisCodeDisplay() {
             </div>
           </div>
         </div>
-        <p className="text-center text-sm text-accent3-500">
+        <p className="text-accent3-500 text-center text-sm">
           Note: If your code doesn't match, restart the game
         </p>
       </div>