+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2025-09-16
+
+### Added
+
+- Added note to the Genesis Code display that it is currently not calculating correctly.
+
 ## 2025-08-14
 
 ### Fixed